# docs/swagger.yaml
{
  "openapi": "3.0.1",
  "info": {
    "title": "Sooktin Swagger",
    "description": "숙틴 v1 명세서입니다",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "bearer ": []
    }
  ],
  "paths": {
    "/users/card-storage/career-cards": {
      "post": {
        "tags": [
          "user-controller"
        ],
        "operationId": "saveCareerCard",
        "parameters": [
          {
            "name": "careerCardId",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoLong"
                }
              }
            }
          }
        }
      }
    },
    "/usernotes": {
      "get": {
        "tags": [
          "usernote-controller"
        ],
        "operationId": "getAllUsernotes",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoListCreateUsernoteResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "usernote-controller"
        ],
        "operationId": "createPost",
        "requestBody": {
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/CreateUsernoteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoCreateUsernoteResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/nickname": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getNickname",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoString"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "user-controller"
        ],
        "operationId": "chacngeNickname",
        "requestBody": {
          "content": {
            "application/json;charset=UTF-8": {
              "schema": {
                "$ref": "#/components/schemas/NicknameRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/NicknameResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/search": {
      "patch": {
        "tags": [
          "search-controller"
        ],
        "operationId": "trackSearchKeyword",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoObject"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/{userId}/main-image": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "getUserMainImage",
        "parameters": [
          {
            "name": "userId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoString"
                }
              }
            }
          }
        }
      }
    },
...
    "/users/search": {
      "get": {
        "tags": [
          "user-controller"
        ],
        "operationId": "searchUsers",
        "parameters": [
          {
            "name": "nickname",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          }
        }
      }
    },
...
    "/usernotes/search": {
      "get": {
        "tags": [
          "usernote-controller"
        ],
        "operationId": "searchUsernote",
        "parameters": [
          {
            "name": "keyword",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pageable",
            "in": "query",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/Pageable"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json;charset=UTF-8": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseDtoSearchUsernoteResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ResponseDtoLong": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "CreateUsernoteRequest": {
        "type": "object",
        "properties": {
          "content": {
            "type": "string"
          }
        }
      },
      "CreateUsernoteResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "userNickname": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "modifiedAt": {
            "type": "string"
          }
        }
      },
      "ResponseDtoCreateUsernoteResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/CreateUsernoteResponse"
          }
        }
      },
...
      "ResponseDtoString": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "NicknameRequest": {
        "required": [
          "nickname"
        ],
        "type": "object",
        "properties": {
          "nickname": {
            "maxLength": 10,
            "minLength": 2,
            "type": "string"
          }
        }
      },
      "NicknameResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string"
          },
          "nickname": {
            "type": "string"
          }
        }
      },
      "ResponseDtoObject": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "CareerCard": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "major": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "modified_at": {
            "type": "string",
            "format": "date-time"
          },
          "imageUrls": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "student_status": {
            "type": "string"
          },
          "grade": {
            "type": "integer",
            "format": "int32"
          },
          "student_num": {
            "type": "string"
          },
          "department": {
            "type": "string"
          },
          "job": {
            "type": "string"
          },
          "experiences": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Experience"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "storageMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StorageCardMapping"
            }
          }
        }
      },
      "CareerCardStorage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "user": {
            "$ref": "#/components/schemas/User"
          },
          "cardMappings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StorageCardMapping"
            }
          }
        }
      },
      "Experience": {
        "type": "object",
        "properties": {
          "company": {
            "type": "string"
          },
          "period": {
            "type": "string"
          }
        }
      },
      "StorageCardMapping": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "careerCard": {
            "$ref": "#/components/schemas/CareerCard"
          },
          "careerCardStorage": {
            "$ref": "#/components/schemas/CareerCardStorage"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "email": {
            "type": "string",
            "description": "email",
            "example": "foo@sookmyung.ac.kr"
          },
          "nickname": {
            "maxLength": 10,
            "minLength": 2,
            "type": "string"
          },
          "password": {
            "type": "string",
            "description": "password",
            "example": "jamone122@3"
          },
          "roles": {
            "uniqueItems": true,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "USER",
                "ADMIN"
              ]
            }
          },
          "careerCard": {
            "$ref": "#/components/schemas/CareerCard"
          },
          "careerCardStorage": {
            "$ref": "#/components/schemas/CareerCardStorage"
          }
        },
        "description": "회원"
      },
...
      "ResponseDtoListCreateUsernoteResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateUsernoteResponse"
            }
          }
        }
      },
...
      "Pageable": {
        "type": "object",
        "properties": {
          "page": {
            "minimum": 0,
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "sort": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "ResponseDtoSearchUsernoteResponse": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/SearchUsernoteResponse"
          }
        }
      },
      "SearchUsernoteResponse": {
        "type": "object",
        "properties": {
          "usernotes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UsernoteDTO"
            }
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UsernoteDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "content": {
            "type": "string"
          },
          "userId": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer ": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}